# Conflicts:
#	back-end/src/main/java/com/example/backend/controller/product_controller/controller/ColorController.java
#	back-end/src/main/java/com/example/backend/controller/product_controller/repository/ColorRepository.java
#	font-end/src/App.js

+++ back-end/src/main/java/com/example/backend/controller/product_controller/service/impl/ColorServiceImpl.java
@@ -25,20 +25,28 @@ public class ColorServiceImpl implements Iservice<Color> {
         return colorRepository.findAll(pageable);
     }
 
-    @Override
-    public void insert(Color color) {
-        colorRepository.save(color);
-
+    public Page<Color> getDelete(Pageable pageable, String key) {
+        return colorRepository.deleteColor(pageable, key);
     }
 
     @Override
-    public void update(Color color, Integer id) {
+    public void insert(Color color) {
         colorRepository.save(color);
+    }
+
+    @Override
+    public void update(Color chip, Integer id) {
+        Color colorUpdate = colorRepository.findById(id).orElse(null);
+        colorUpdate.setCode(chip.getCode());
+        colorUpdate.setName(chip.getName());
+        colorRepository.save(colorUpdate);
     }
 
     @Override
     public void delete(Integer id) {
-        colorRepository.deleteById(id);
+        if (colorRepository.existsById(id)){
+            colorRepository.deleteById(id);
+        }
     }
 
     @Override
@@ -57,7 +65,6 @@ public class ColorServiceImpl implements Iservice<Color> {
         return colorRepository.findById(id).get();
     }
 
-
     public void importDataFromExcel(MultipartFile file) throws Exception{
         InputStream inputStream = file.getInputStream();
         Workbook workbook = new XSSFWorkbook(inputStream);
@@ -67,10 +74,8 @@ public class ColorServiceImpl implements Iservice<Color> {
             if(row.getRowNum() == 0){
                 continue;
             }
-
             String code = row.getCell(0).getStringCellValue();
             Color existingColor = colorRepository.findByCode(code);
-
             if(existingColor != null){
                 //Đã tồn tại
                 existingColor.setName(row.getCell(1).getStringCellValue());
@@ -94,8 +99,8 @@ public class ColorServiceImpl implements Iservice<Color> {
         workbook.close();
     }
 
-    public Page<Color> search(String search,Pageable pageable) {
-        return colorRepository.search(search ,pageable);
+    public Page<Color> search(Pageable pageable, String key) {
+        return colorRepository.search(pageable, key);
     }
 
 }
